Added the Final Project Report
Note: The Final Report uploaded here is a Reduced version where the number of POSTMAN test result screenshots attached do not match all possible test cases due to concerns of the file sizes. The attached images demonstrate the basics of accessing the API endpoints of each web service whereas all of the endpoints are not mentioned in the documentation due to the file size concerns.